+++ packages/dockerfile.6.4.0/src-opam/dockerfile_distro.ml
@@ -19,13 +19,13 @@
 open Astring
 
 type t = [
-  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `Latest ]
-  | `CentOS of [ `V6 | `V7 | `Latest ]
+  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `Latest ]
+  | `CentOS of [ `V6 | `V7 | `V8 | `Latest ]
   | `Debian of [ `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `Latest ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `Latest ]
   | `OracleLinux of [ `V7 | `Latest ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `Latest ]
-  | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `LTS | `Latest ]
+  | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `V19_10 | `V20_04 | `LTS | `Latest ]
 ] [@@deriving sexp]
 
 type status = [
@@ -35,23 +35,23 @@ type status = [
 ] [@@deriving sexp]
 
 let distros = [
-  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `Latest;
-  `CentOS `V6; `CentOS `V7; `CentOS `Latest;
+  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `V3_11; `Alpine `Latest;
+  `CentOS `V6; `CentOS `V7; `CentOS `V8; `CentOS `Latest;
   `Debian `V10; `Debian `V9; `Debian `V8; `Debian `V7;
   `Debian `Stable; `Debian `Testing; `Debian `Unstable;
-  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `Latest;
+  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `V31; `Fedora `Latest;
   `OracleLinux `V7; `OracleLinux `Latest;
   `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `V15_1; `OpenSUSE `Latest;
   `Ubuntu `V12_04; `Ubuntu `V14_04; `Ubuntu `V15_04; `Ubuntu `V15_10;
-  `Ubuntu `V16_04; `Ubuntu `V16_10; `Ubuntu `V17_04; `Ubuntu `V17_10; `Ubuntu `V18_04; `Ubuntu `V18_10; `Ubuntu `V19_04;
+  `Ubuntu `V16_04; `Ubuntu `V16_10; `Ubuntu `V17_04; `Ubuntu `V17_10; `Ubuntu `V18_04; `Ubuntu `V18_10; `Ubuntu `V19_04; `Ubuntu `V19_10; `Ubuntu `V20_04;
   `Ubuntu `Latest; `Ubuntu `LTS ]
 
 let distro_status (d:t) : status = match d with
-  | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8) -> `Deprecated
-  | `Alpine `V3_9 -> `Active `Tier2
-  | `Alpine `V3_10 -> `Active `Tier1
-  | `Alpine `Latest -> `Alias (`Alpine `V3_10)
-  | `CentOS `V7 -> `Active `Tier2
+  | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9) -> `Deprecated
+  | `Alpine `V3_10 -> `Active `Tier2
+  | `Alpine `V3_11 -> `Active `Tier1
+  | `Alpine `Latest -> `Alias (`Alpine `V3_11)
+  | `CentOS (`V7 | `V8) -> `Active `Tier2
   | `CentOS `V6 -> `Deprecated
   | `CentOS `Latest -> `Alias (`CentOS `V7)
   | `Debian `V7 -> `Deprecated
@@ -61,18 +61,18 @@ let distro_status (d:t) : status = match d with
   | `Debian `Stable -> `Alias (`Debian `V10)
   | `Debian `Testing -> `Active `Tier2
   | `Debian `Unstable -> `Active `Tier2
-  | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29) -> `Deprecated
-  | `Fedora `V30 -> `Active `Tier2
-  | `Fedora `Latest -> `Alias (`Fedora `V30)
+  | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30) -> `Deprecated
+  | `Fedora `V31 -> `Active `Tier2
+  | `Fedora `Latest -> `Alias (`Fedora `V31)
   | `OracleLinux `V7 -> `Active `Tier2
   | `OracleLinux `Latest -> `Alias (`OracleLinux `V7)
   | `OpenSUSE (`V42_1 | `V42_2 | `V42_3 | `V15_0) -> `Deprecated
   | `OpenSUSE `V15_1 -> `Active `Tier2
   | `OpenSUSE `Latest -> `Alias (`OpenSUSE `V15_1)
-  | `Ubuntu (`V16_04 | `V18_04 | `V19_04 ) -> `Active `Tier2
-  | `Ubuntu ( `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_10 | `V17_04 | `V17_10 | `V18_10) -> `Deprecated
+  | `Ubuntu (`V16_04 | `V18_04 | `V19_10 | `V20_04) -> `Active `Tier2
+  | `Ubuntu ( `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_10 | `V17_04 | `V17_10 | `V18_10 | `V19_04) -> `Deprecated
   | `Ubuntu `LTS -> `Alias (`Ubuntu `V18_04)
-  | `Ubuntu `Latest -> `Alias (`Ubuntu `V19_04)
+  | `Ubuntu `Latest -> `Alias (`Ubuntu `V19_10)
 
 let latest_distros =
   [ `Alpine `Latest; `CentOS `Latest;
@@ -91,8 +91,8 @@ module OV = Ocaml_version
 let distro_arches ov (d:t) =
   match resolve_alias d, ov with
   | `Debian (`V9|`V10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
-  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
-  | `Ubuntu (`V16_04|`V18_04|`V17_10|`V18_10|`V19_04), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
+  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
+  | `Ubuntu (`V16_04|`V18_04|`V17_10|`V18_10|`V19_04|`V19_10|`V20_04), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
   | _ -> [ `X86_64 ]
 
 
@@ -129,6 +129,8 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Ubuntu `V18_04 -> Some "4.05.0"
   |`Ubuntu `V18_10 -> Some "4.05.0"
   |`Ubuntu `V19_04 -> Some "4.05.0"
+  |`Ubuntu `V19_10 -> Some "4.05.0"
+  |`Ubuntu `V20_04 -> Some "4.08.1"
   |`Alpine `V3_3 -> Some "4.02.3"
   |`Alpine `V3_4 -> Some "4.02.3"
   |`Alpine `V3_5 -> Some "4.04.0"
@@ -137,6 +139,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Alpine `V3_8 -> Some "4.06.1"
   |`Alpine `V3_9 -> Some "4.06.1"
   |`Alpine `V3_10 -> Some "4.07.0"
+  |`Alpine `V3_11 -> Some "4.08.1"
   |`Fedora `V21 -> Some "4.01.0"
   |`Fedora `V22 -> Some "4.02.0"
   |`Fedora `V23 -> Some "4.02.2"
@@ -147,8 +150,10 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Fedora `V28 -> Some "4.06.0"
   |`Fedora `V29 -> Some "4.07.0"
   |`Fedora `V30 -> Some "4.07.0"
+  |`Fedora `V31 -> Some "4.08.1"
   |`CentOS `V6 -> Some "3.11.2"
   |`CentOS `V7 -> Some "4.01.0"
+  |`CentOS `V8 -> Some "4.07.0"
   |`OpenSUSE `V42_1 -> Some "4.02.3"
   |`OpenSUSE `V42_2 -> Some "4.03.0"
   |`OpenSUSE `V42_3 -> Some "4.03.0"
@@ -172,6 +177,8 @@ let tag_of_distro (d:t) = match d with
   |`Ubuntu `V18_04 -> "ubuntu-18.04"
   |`Ubuntu `V18_10 -> "ubuntu-18.10"
   |`Ubuntu `V19_04 -> "ubuntu-19.04"
+  |`Ubuntu `V19_10 -> "ubuntu-19.10"
+  |`Ubuntu `V20_04 -> "ubuntu-20.04"
   |`Ubuntu `Latest -> "ubuntu"
   |`Ubuntu `LTS -> "ubuntu-lts"
   |`Debian `Stable -> "debian-stable"
@@ -183,6 +190,7 @@ let tag_of_distro (d:t) = match d with
   |`Debian `V7 -> "debian-7"
   |`CentOS `V6 -> "centos-6"
   |`CentOS `V7 -> "centos-7"
+  |`CentOS `V8 -> "centos-8"
   |`CentOS `Latest -> "centos"
   |`Fedora `Latest -> "fedora"
   |`Fedora `V21 -> "fedora-21"
@@ -195,6 +203,7 @@ let tag_of_distro (d:t) = match d with
   |`Fedora `V28 -> "fedora-28"
   |`Fedora `V29 -> "fedora-29"
   |`Fedora `V30 -> "fedora-30"
+  |`Fedora `V31 -> "fedora-31"
   |`OracleLinux `V7 -> "oraclelinux-7"
   |`OracleLinux `Latest -> "oraclelinux"
   |`Alpine `V3_3 -> "alpine-3.3"
@@ -205,6 +214,7 @@ let tag_of_distro (d:t) = match d with
   |`Alpine `V3_8 -> "alpine-3.8"
   |`Alpine `V3_9 -> "alpine-3.9"
   |`Alpine `V3_10 -> "alpine-3.10"
+  |`Alpine `V3_11 -> "alpine-3.11"
   |`Alpine `Latest -> "alpine"
   |`OpenSUSE `V42_1 -> "opensuse-42.1"
   |`OpenSUSE `V42_2 -> "opensuse-42.2"
@@ -225,6 +235,8 @@ let distro_of_tag x : t option = match x with
   |"ubuntu-18.04" -> Some (`Ubuntu `V18_04)
   |"ubuntu-18.10" -> Some (`Ubuntu `V18_10)
   |"ubuntu-19.04" -> Some (`Ubuntu `V19_04)
+  |"ubuntu-19.10" -> Some (`Ubuntu `V19_10)
+  |"ubuntu-20.04" -> Some (`Ubuntu `V20_04)
   |"ubuntu" -> Some (`Ubuntu `Latest)
   |"ubuntu-lts" -> Some (`Ubuntu `LTS)
   |"debian-stable" -> Some (`Debian `Stable)
@@ -236,6 +248,7 @@ let distro_of_tag x : t option = match x with
   |"debian-7" -> Some (`Debian `V7)
   |"centos-6" -> Some (`CentOS `V6)
   |"centos-7" -> Some (`CentOS `V7)
+  |"centos-8" -> Some (`CentOS `V8)
   |"fedora-21" -> Some (`Fedora `V21)
   |"fedora-22" -> Some (`Fedora `V22)
   |"fedora-23" -> Some (`Fedora `V23)
@@ -246,6 +259,7 @@ let distro_of_tag x : t option = match x with
   |"fedora-28" -> Some (`Fedora `V28)
   |"fedora-29" -> Some (`Fedora `V29)
   |"fedora-30" -> Some (`Fedora `V30)
+  |"fedora-31" -> Some (`Fedora `V31)
   |"fedora" -> Some (`Fedora `Latest)
   |"oraclelinux-7" -> Some (`OracleLinux `V7)
   |"oraclelinux" -> Some (`OracleLinux `Latest)
@@ -257,6 +271,7 @@ let distro_of_tag x : t option = match x with
   |"alpine-3.8" -> Some (`Alpine `V3_8)
   |"alpine-3.9" -> Some (`Alpine `V3_9)
   |"alpine-3.10" -> Some (`Alpine `V3_10)
+  |"alpine-3.11" -> Some (`Alpine `V3_11)
   |"alpine" -> Some (`Alpine `Latest)
   |"opensuse-42.1" -> Some (`OpenSUSE `V42_1)
   |"opensuse-42.2" -> Some (`OpenSUSE `V42_2)
@@ -280,6 +295,8 @@ let rec human_readable_string_of_distro (d:t) =
   |`Ubuntu `V18_04 -> "Ubuntu 18.04"
   |`Ubuntu `V18_10 -> "Ubuntu 18.10"
   |`Ubuntu `V19_04 -> "Ubuntu 19.04"
+  |`Ubuntu `V19_10 -> "Ubuntu 19.10"
+  |`Ubuntu `V20_04 -> "Ubuntu 20.04"
   |`Debian `Stable -> "Debian Stable"
   |`Debian `Unstable -> "Debian Unstable"
   |`Debian `Testing -> "Debian Testing"
@@ -289,6 +306,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Debian `V7 -> "Debian 7 (Wheezy)"
   |`CentOS `V6 -> "CentOS 6"
   |`CentOS `V7 -> "CentOS 7"
+  |`CentOS `V8 -> "CentOS 8"
   |`Fedora `V21 -> "Fedora 21"
   |`Fedora `V22 -> "Fedora 22"
   |`Fedora `V23 -> "Fedora 23"
@@ -299,6 +317,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Fedora `V28 -> "Fedora 28"
   |`Fedora `V29 -> "Fedora 29"
   |`Fedora `V30 -> "Fedora 30"
+  |`Fedora `V31 -> "Fedora 31"
   |`OracleLinux `V7 -> "OracleLinux 7"
   |`Alpine `V3_3 -> "Alpine 3.3"
   |`Alpine `V3_4 -> "Alpine 3.4"
@@ -308,6 +327,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Alpine `V3_8 -> "Alpine 3.8"
   |`Alpine `V3_9 -> "Alpine 3.9"
   |`Alpine `V3_10 -> "Alpine 3.10"
+  |`Alpine `V3_11 -> "Alpine 3.11"
   |`OpenSUSE `V42_1 -> "OpenSUSE 42.1"
   |`OpenSUSE `V42_2 -> "OpenSUSE 42.2"
   |`OpenSUSE `V42_3 -> "OpenSUSE 42.3"
@@ -362,6 +382,7 @@ let base_distro_tag d =
           | `V3_8 -> "3.8"
           | `V3_9 -> "3.9"
           | `V3_10 -> "3.10"
+          | `V3_11 -> "3.11"
           | `Latest -> assert false
         in
         "alpine", tag
@@ -391,11 +412,13 @@ let base_distro_tag d =
           | `V18_04 -> "bionic"
           | `V18_10 -> "cosmic"
           | `V19_04 -> "disco"
+          | `V19_10 -> "eoan"
+          | `V20_04 -> "focal"
           | `Latest | `LTS -> assert false
         in
         "ubuntu", tag
     | `CentOS v ->
-        let tag = match v with `V6 -> "6" | `V7 -> "7" | _ -> assert false in
+        let tag = match v with `V6 -> "6" | `V7 -> "7" | `V8 -> "8" | _ -> assert false in
         "centos", tag
     | `Fedora v ->
         let tag =
@@ -410,6 +433,7 @@ let base_distro_tag d =
           | `V28 -> "28"
           | `V29 -> "29"
           | `V30 -> "30"
+          | `V31 -> "31"
           | `Latest -> assert false
         in
         "fedora", tag