+++ packages/dockerfile.6.4.0/src-opam/dockerfile_opam.ml
@@ -78,6 +78,7 @@ let apk_opam2 ?(labels= []) ~distro ~tag () =
   @@ install_opam_from_source ~branch:"2.0" ()
   @@ run "strip /usr/local/bin/opam*"
   @@ from ~tag distro
+  @@ Linux.Apk.add_repository ~tag:"testing" "http://dl-cdn.alpinelinux.org/alpine/edge/testing"
   @@ copy ~from:"0" ~src:["/usr/local/bin/opam"] ~dst:"/usr/bin/opam" ()
   @@ copy ~from:"0" ~src:["/usr/local/bin/opam-installer"]
        ~dst:"/usr/bin/opam-installer" ()
@@ -85,8 +86,6 @@ let apk_opam2 ?(labels= []) ~distro ~tag () =
   @@ Linux.Apk.dev_packages ()
   @@ Linux.Apk.add_user ~uid:1000 ~gid:1000 ~sudo:true "opam"
   @@ install_bubblewrap_wrappers @@ Linux.Git.init ()
-  @@ run
-       "git clone git://github.com/ocaml/opam-repository /home/opam/opam-repository"
 
 
 (* Debian based Dockerfile *)
@@ -104,20 +103,26 @@ let apt_opam2 ?(labels= []) ~distro ~tag () =
   @@ Linux.Apt.dev_packages ()
   @@ Linux.Apt.add_user ~uid:1000 ~sudo:true "opam"
   @@ install_bubblewrap_wrappers @@ Linux.Git.init ()
-  @@ run
-       "git clone git://github.com/ocaml/opam-repository /home/opam/opam-repository"
 
 
-(* RPM based Dockerfile *)
-let yum_opam2 ?(labels= []) ~distro ~tag () =
+(* RPM based Dockerfile.
+   [yum_workaround activates the overlay/yum workaround needed
+   for older versions of yum as found in CentOS 7 and earlier *)
+let yum_opam2 ?(labels= []) ~yum_workaround ~distro ~tag () =
+  let workaround =
+    if yum_workaround then
+      run "touch /var/lib/rpm/*"
+      @@ Linux.RPM.install "yum-plugin-ovl"
+    else empty
+  in
   header distro tag @@ label (("distro_style", "rpm") :: labels)
-  @@ run "touch /var/lib/rpm/*"
-  @@ Linux.RPM.install "yum-plugin-ovl"
-  @@ Linux.RPM.update 
+  @@ workaround
+  @@ Linux.RPM.update
   @@ Linux.RPM.dev_packages ~extra:"which tar curl xz libcap-devel openssl" ()
   @@ install_bubblewrap_from_source ()
   @@ install_opam_from_source ~prefix:"/usr" ~branch:"2.0" ()
-  @@ from ~tag distro @@ Linux.RPM.install "yum-plugin-ovl" @@ Linux.RPM.update
+  @@ from ~tag distro @@ workaround
+  @@ Linux.RPM.update
   @@ Linux.RPM.dev_packages ()
   @@ copy ~from:"0" ~src:["/usr/local/bin/bwrap"] ~dst:"/usr/bin/bwrap" ()
   @@ copy ~from:"0" ~src:["/usr/bin/opam"] ~dst:"/usr/bin/opam" ()
@@ -127,8 +132,6 @@ let yum_opam2 ?(labels= []) ~distro ~tag () =
        "sed -i.bak '/LC_TIME LC_ALL LANGUAGE/aDefaults    env_keep += \"OPAMYES OPAMJOBS OPAMVERBOSE\"' /etc/sudoers"
   @@ Linux.RPM.add_user ~uid:1000 ~sudo:true "opam"
   @@ install_bubblewrap_wrappers @@ Linux.Git.init ()
-  @@ run
-       "git clone git://github.com/ocaml/opam-repository /home/opam/opam-repository"
 
 
 (* Zypper based Dockerfile *)
@@ -145,17 +148,21 @@ let zypper_opam2 ?(labels= []) ~distro ~tag () =
        ~dst:"/usr/bin/opam-installer" ()
   @@ Linux.Zypper.add_user ~uid:1000 ~sudo:true "opam"
   @@ install_bubblewrap_wrappers @@ Linux.Git.init ()
-  @@ run
-       "git clone git://github.com/ocaml/opam-repository /home/opam/opam-repository"
 
-let gen_opam2_distro ?labels d =
+let gen_opam2_distro ?(clone_opam_repo=true) ?labels d =
   let distro, tag = D.base_distro_tag d in
   let fn = match D.package_manager d with
   | `Apk -> apk_opam2 ?labels ~tag ~distro ()
   | `Apt -> apt_opam2 ?labels ~tag ~distro ()
-  | `Yum -> yum_opam2 ?labels ~tag ~distro ()
+  | `Yum ->
+     let yum_workaround = match d with `CentOS `V7 -> true | _ -> false in
+     yum_opam2 ?labels ~yum_workaround ~tag ~distro ()
   | `Zypper -> zypper_opam2 ?labels ~tag ~distro ()
-  in (D.tag_of_distro d, fn)
+  in
+  let clone = if clone_opam_repo then
+    run "git clone git://github.com/ocaml/opam-repository /home/opam/opam-repository"
+  else empty in
+  (D.tag_of_distro d, fn @@ clone)
 
 (* Generate archive mirror *)
 let opam2_mirror (hub_id: string) =